% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{annual_exceedances}
\alias{annual_exceedances}
\title{Calculate Annual Exceedances}
\usage{
annual_exceedances(tidy_df, station, start_date, end_date,
  april_start = FALSE)
}
\arguments{
\item{tidy_df}{a df}
}
\value{
a data frame of results
}
\description{
Calculates the number of annual exceedances
}
