% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{as.result}
\alias{as.result}
\title{Clean Result}
\usage{
as.result(x)
}
\arguments{
\item{x}{a numeric or character vector}
}
\value{
a numeric vector
}
\description{
Clean a character vector of results and convert into a numeric vector.
}
