% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{as.sampledate}
\alias{as.sampledate}
\title{As SampleDate}
\usage{
as.sampledate(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a vector of dates
}
\description{
Convert a character vector with dates into a vector of elements in the date format.
}
