% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{bact_check}
\alias{bact_check}
\title{Analyze Multiple Stations}
\usage{
bact_check(df, sites, BU, water_type, ...)
}
\arguments{
\item{df}{a data frame}

\item{sites}{a character vector with names of the sites}

\item{BU}{a character vector with the highest beneficial use for the associated site}

\item{water_type}{a string of either "fresh" or "marine' representing the water type for the site}
}
\value{
a list of data frames that correspond to each site
}
\description{
Runs bacteria checker on multiple stations
}
