% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{bact_geomeans}
\alias{bact_geomeans}
\title{Calculate Geometric Means}
\usage{
bact_geomeans(df)
}
\arguments{
\item{df}{a data frame of indicator bacteria monitoring data that has been tidied by
tidy_bacteria(df) and has daily rows between the first and last SampleDate using expand_dates(df)}
}
\value{
an expanded data frame with columns for geometric means and sample counts
}
\description{
Expands the input data frame by adding columns for 42-day rolling geometric means and counts
of samples used to calculate the geometric mean.
}
