% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{check_fecal_to_total}
\alias{check_fecal_to_total}
\title{Check Fecal-To-Total Ratio}
\usage{
check_fecal_to_total(df, sub_ecoli_for_fecal = FALSE, water_type = "marine")
}
\arguments{
\item{df}{a tidy data frame}

\item{water_type}{a string of either "fresh" or "marine' representing the water type
of the water body}

\item{sub_ecoli_for_frecal}{a logical object that is TRUE if e. coli results should be used in
place of fecal coliform to caculate the fecal to total ratio}
}
\value{
a data frame
}
\description{
Inserts columns that calculate the fecal coliform to total coliform ratio, show the single sample
objective for total coliform if applicable, and show compliance with applicable objectives.
}
