% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{check_geolimits}
\alias{check_geolimits}
\title{Check Geometric Means Against Limitations}
\usage{
check_geolimits(df, BU = "REC-1", water_type = "marine")
}
\arguments{
\item{df}{a data frame of indicator bacteria monitoring data that has been tidied by
tidy_bacteria(df), has daily rows between the first and last SampleDate using expand_dates(df)
and includes geometric mean calculations using bact_geomeans(df)}
}
\value{
an expanded data frame with columns for relevant water quality objectives and compliance
}
\description{
Expands the input data frame to include columns that list relevant geometric mean water quality
objectives and compliance with the objective.
}
