% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{check_sslimits}
\alias{check_sslimits}
\title{Check Single Sample Results Against Limitations}
\usage{
check_sslimits(df, BU = "REC-1", water_type = "marine")
}
\arguments{
\item{df}{a data frame of indicator bacteria monitoring data that has been tidied by
tidy_bacteria(df)}
}
\value{
an expanded data frame with columns for relevant water quality objectives and compliance
}
\description{
Expands the input data frame to include columns that list relevant single sample water quality
objectives and compliance with the objective.
}
