% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{col_check}
\alias{col_check}
\title{Column Headers Check}
\usage{
col_check(df, x)
}
\arguments{
\item{df}{a data frame of water quality data}

\item{x}{a character vector of column names}
}
\value{
a logical value that is TRUE if all elements of names are columns in the data frame
}
\description{
Checks if a data frame includes specified column names.
}
\examples{
col_check(df, c("StationCode", "SampleDate", "WeatherCondition"))
col_check(df, required_columns)
}
