% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{column_clean}
\alias{column_clean}
\title{Clean Column Type}
\usage{
column_clean(df)
}
\arguments{
\item{df}{a data frame of monitoring data with CEDEN-named columns}
}
\value{
a data frame
}
\description{
Corrects the column data types for standard columns.
}
