% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{convertWeather}
\alias{convertWeather}
\title{Convert Weather Condition}
\usage{
convertWeather(df)
}
\arguments{
\item{df}{a data frame with WeatherCondition and SampleDate columns}
}
\value{
a data frame
}
\description{
Convert WeatherCondition entries of "Dry" to "Winter Dry" and "Summer Dry" depending on the date.
"Winter Dry" are dry weather days occuring on November 1 through March 31. "Summer Dry" are dry
weather days occurring on April 1 through October 31.
}
