% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{expand_dates}
\alias{expand_dates}
\title{Insert Daily Rows}
\usage{
expand_dates(df)
}
\arguments{
\item{df}{a data frame of monitoring data that has been tidied by tidy_bacteria}
}
\value{
a data frame
}
\description{
Inserts a row for each day between the first and last dates found in the SampleDate
column of a data frame.
}
