% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{first_date}
\alias{first_date}
\title{Get First Date}
\usage{
first_date(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a date
}
\description{
Returns the first date found in the SampleDate column of a data frame.
}
