% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{grab_limits_geo}
\alias{grab_limits_geo}
\title{Grab Geomteric Mean Limitations}
\usage{
grab_limits_geo(BU, water_type)
}
\arguments{
\item{BU}{a string of either "REC-1", "LREC-1", or "REC-2" representing the highest
beneficial use of the water body}

\item{water_type}{a string of either "fresh" or "marine' representing the water type
of the water body}
}
\value{
a numeric vector with the geometric mean objectives for e. coli, fecal coliform,
total coliform, and enterococcus
}
\description{
Retrieves geometric mean water quality objectives for indicator bacteria.
}
