% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{last_date}
\alias{last_date}
\title{Get Last Date}
\usage{
last_date(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a date
}
\description{
Returns the last date found in the SampleDate column of a data frame.
}
