% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{replace_nd}
\alias{replace_nd}
\title{Replace Non Detects}
\usage{
replace_nd(df, assume_mdl = TRUE, ecoli_mdl = 100, fc_mdl = 100,
  tc_mdl = 100, ent_mdl = 10)
}
\arguments{
\item{df}{a data frame of monitoring data that has been tidied by tidy_bacteria}

\item{assume_mdl}{a logical}

\item{ecoli_mdl}{a numeric}

\item{fc_mdl}{a numeric}

\item{tc_mdl}{a numeric}

\item{ent_mdl}{a numeric}
}
\value{
a data frame
}
\description{
Replace non detects in the data to allow for the calculation of geometric means.
}
