% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{tidy_bacteria}
\alias{tidy_bacteria}
\title{Tidy Bacteria Data}
\usage{
tidy_bacteria(df)
}
\arguments{
\item{df}{a data frame of monitoring data.}
}
\value{
a data frame of monitoring data that selects indicator bacteria results and reorganizes
results by sampling date.
}
\description{
Tidies water quality data so that bacteria data is grouped together.
}
\section{Warning}{

For this code to work, the input data frame should in an analytical record format
(e.g. CEDEN) that includes columns for the following:

StationCode, SampleDate, WeatherCode, AnalyteName, Result, ResQualCode, MDL, and RL

Indiciator bacteria names in the "AnalyteName" field should be labeled as followed:

"E. coli", "Enterococcus", "Coliform, Fecal", and "Coliform, Total"
}

\examples{
tidy_bacteria(smb_beachdata_15-16)
tidy_bacteria(CEDEN_export)
}
